+++ src/type_props.rs
@@ -5,8 +5,11 @@ use core::{
     mem::{align_of, align_of_val, size_of, size_of_val},
     ptr::NonNull,
 };
+use crate::helpers::dangling_nonnull;
 
 /// The maximum value of a `usize` with no high bit.
+///
+/// Equivalent to `usize::MAX >> 1` or `isize::MAX as usize`.
 pub const USIZE_MAX_NO_HIGH_BIT: usize = usize::MAX >> 1;
 
 /// A trait containing constants for sized types.
@@ -37,7 +40,7 @@ pub trait PtrProps<T: ?Sized> {
     /// # Safety
     ///
     /// The pointer must be valid as a reference.
-    unsafe fn size(&self) -> usize;
+	unsafe fn size(&self) -> usize;
     /// Gets the alignment of the value.
     ///
     /// # Safety
@@ -49,7 +52,10 @@ pub trait PtrProps<T: ?Sized> {
     /// # Safety
     ///
     /// The pointer must be valid as a reference.
-    unsafe fn layout(&self) -> Layout;
+    #[inline]
+    unsafe fn layout(&self) -> Layout {
+        Layout::from_size_align_unchecked(self.size(), self.align())
+    }
 
     #[cfg(feature = "metadata")]
     /// Gets the metadata of the value.
@@ -82,51 +88,91 @@ pub trait PtrProps<T: ?Sized> {
     }
 }
 
-/// Implements [`PtrProps`] for a pointer type.
-macro_rules! impl_ptr_props {
-	($($name:ty $(,$to_ptr:ident)?)*) => {
-		$(
-		impl<T: ?Sized> PtrProps<T> for $name {
-			unsafe fn size(&self) -> usize {
-				// We use &*(*val) (?.to_ptr()?) to convert any primitive pointer type to a
-                //  reference.
-				// This is kind of a hack, but it lets us avoid *_of_val_raw, which is unstable.
-				size_of_val::<T>(&*(*self)$(.$to_ptr())?)
-			}
-
-			unsafe fn align(&self) -> usize {
-				align_of_val::<T>(&*(*self)$(.$to_ptr())?)
-			}
-
-			unsafe fn layout(&self) -> Layout {
-				Layout::from_size_align_unchecked(
-					self.size(),
-					self.align()
-				)
-			}
-
-			#[cfg(feature = "metadata")]
-			unsafe fn metadata(&self) -> <T as core::ptr::Pointee>::Metadata {
-				core::ptr::metadata(&*(*self)$(.$to_ptr())?)
-			}
-		}
-		)*
-	}
+macro_rules! impl_ptr_props_raw {
+    ($($name:ty),* $(,)?) => {
+        $(
+            impl<T: ?Sized> PtrProps<T> for $name {
+                #[inline]
+                unsafe fn size(&self) -> usize {
+                    size_of_val::<T>(&**self)
+                }
+                #[inline]
+                unsafe fn align(&self) -> usize {
+                    align_of_val::<T>(&**self)
+                }
+                #[cfg(feature = "metadata")]
+                unsafe fn metadata(&self) -> <T as core::ptr::Pointee>::Metadata {
+                    core::ptr::metadata(&*(*self))
+                }
+            }
+        )*
+    };
 }
 
-impl_ptr_props!(
-    *const T
-    *mut T
+macro_rules! impl_ptr_props_identity {
+    ($($name:ty),* $(,)?) => {
+        $(
+            impl<T: ?Sized> PtrProps<T> for $name {
+                #[inline]
+                unsafe fn size(&self) -> usize {
+                    size_of_val::<T>(*self)
+                }
+                #[inline]
+                unsafe fn align(&self) -> usize {
+                    align_of_val::<T>(*self)
+                }
+                #[cfg(feature = "metadata")]
+                unsafe fn metadata(&self) -> <T as core::ptr::Pointee>::Metadata {
+                    core::ptr::metadata(*self)
+                }
+            }
+        )*
+    };
+}
 
-    &T
-    &mut T
+macro_rules! impl_ptr_props_as_ref {
+    ($($name:ty),* $(,)?) => {
+        $(
+            impl<T: ?Sized> PtrProps<T> for $name {
+                #[inline]
+                unsafe fn size(&self) -> usize {
+                    size_of_val::<T>(self.as_ref())
+                }
+                #[inline]
+                unsafe fn align(&self) -> usize {
+                    align_of_val::<T>(self.as_ref())
+                }
+                #[cfg(feature = "metadata")]
+                unsafe fn metadata(&self) -> <T as core::ptr::Pointee>::Metadata {
+                    core::ptr::metadata(self.as_ref())
+                }
+            }
+        )*
+    };
+}
 
-    NonNull<T>, as_ptr
-
-    alloc::boxed::Box<T>
-    alloc::rc::Rc<T>
+impl_ptr_props_raw! { *const T, *mut T }
+impl_ptr_props_identity! { &T, &mut T }
+impl_ptr_props_as_ref! {
+    alloc::boxed::Box<T>,
+    alloc::rc::Rc<T>,
     alloc::sync::Arc<T>
-);
+}
+
+impl<T: ?Sized> PtrProps<T> for NonNull<T> {
+	#[inline]
+	unsafe fn size(&self) -> usize {
+		size_of_val::<T>(&*self.as_ptr())
+	}
+	#[inline]
+	unsafe fn align(&self) -> usize {
+		align_of_val::<T>(&*self.as_ptr())
+	}
+	#[cfg(feature = "metadata")]
+	unsafe fn metadata(&self) -> <T as core::ptr::Pointee>::Metadata {
+		core::ptr::metadata(&*self.as_ptr())
+	}
+}
 
 #[cfg(not(feature = "metadata"))]
 /// Trait for unsized types whose metadata is `usize` (e.g., slices, `str`).
@@ -152,6 +198,10 @@ pub unsafe trait VarSized: core::ptr::Pointee<Metadata = usize> {
     const ALIGN: usize;
 }
 
+unsafe impl<T> VarSized for [T] {
+    const ALIGN: usize = T::ALIGN;
+}
+
 unsafe impl VarSized for str {
     const ALIGN: usize = u8::ALIGN;
 }
@@ -161,7 +211,7 @@ unsafe impl VarSized for core::ffi::CStr {
     const ALIGN: usize = u8::ALIGN;
 }
 #[cfg(feature = "std")]
-// `OsStr == [u8]` and `[u8]: UnsizedCopy`
+// `OsStr == [u8]` and `[u8]: VarSized`
 unsafe impl VarSized for std::ffi::OsStr {
     const ALIGN: usize = u8::ALIGN;
 }
@@ -171,16 +221,39 @@ unsafe impl VarSized for std::path::Path {
     const ALIGN: usize = u8::ALIGN;
 }
 
-unsafe impl<T> VarSized for [T] {
-    const ALIGN: usize = T::ALIGN;
-}
-
 // not associated to reduce clutter, and so they can be const
 
-#[cfg(feature = "metadata")]
 /// Creates a dangling, zero-length, [`NonNull`] pointer with the proper alignment.
 #[must_use]
 pub const fn varsized_dangling_nonnull<T: ?Sized + VarSized>() -> NonNull<T> {
-    // SAFETY: `ALIGN` is guaranteed to be a valid alignment for `Self`.
-    unsafe { NonNull::from_raw_parts(crate::helpers::dangling_nonnull(T::ALIGN), 0) }
+    varsized_nonnull_from_raw_parts(unsafe { dangling_nonnull(T::ALIGN) }, 0)
+}
+
+#[must_use]
+pub const fn varsized_dangling_pointer<T: ?Sized + VarSized>() -> *mut T {
+    varsized_pointer_from_raw_parts(unsafe { dangling_nonnull(T::ALIGN).as_ptr() }, 0)
+}
+
+/// Creates a `NonNull<T>` from a pointer and a `usize` size metadata.
+#[must_use]
+#[inline]
+pub const fn varsized_nonnull_from_raw_parts<T: ?Sized + VarSized>(
+    p: NonNull<u8>,
+    meta: usize,
+) -> NonNull<T> {
+    unsafe { NonNull::new_unchecked(varsized_pointer_from_raw_parts(p.as_ptr(), meta)) }
+}
+
+/// Creates a `*mut T` from a pointer and a `usize` size metadata.
+#[must_use]
+#[inline]
+pub const fn varsized_pointer_from_raw_parts<T: ?Sized + VarSized>(
+    p: *mut u8,
+    meta: usize,
+) -> *mut T {
+    // SAFETY: VarSized trait requires T::Metadata == usize
+    unsafe {
+        // i hate this so much
+        *((&(p, meta)) as *const (*mut u8, usize)).cast::<*mut T>()
+    }
 }