+++ tests/potential_ub.rs
@@ -1,37 +1,68 @@
-use std::ptr::NonNull;
 // these tests are tests which check for potential UB, such as manually constructing a pointer with
 //  metadata via transmutation.
+use core::ptr::NonNull;
 use memapi::helpers::{nonnull_slice_from_raw_parts, nonnull_slice_len, slice_ptr_from_raw_parts};
 
 #[test]
-fn slice_ptr_from_parts_works() {
-    let slice: &mut [usize] = &mut [64, 128, 256, 512];
+fn slice_pointer_from_raw_parts_works() {
+    match sp_frp_inner() {
+        // 0 == success, no ub
+        0 => (),
+        1 => panic!("slice_ptr_from_raw_parts() test 1 failed: result doesn't dereference properly"),
+        2 => panic!("slice_ptr_from_raw_parts() test 2 failed: result doesn't have the same pointer"),
+        3 => panic!("slice_ptr_from_raw_parts() test 3 failed: result doesn't have the same length"),
+        4 => panic!("slice_ptr_from_raw_parts() test 4 failed: result as nonnull doesn't have the same length"),
+        5 => panic!("slice_ptr_from_raw_parts() test 5 failed: result doesn't have the same values"),
+        6 => panic!("slice_ptr_from_raw_parts() test 6 failed: result doesn't have the correct values"),
+        _ => unreachable!(),
+    }
+}
+
+fn sp_frp_inner() -> usize {
+    let i = 4;
+
+    let mut data: Vec<usize> = (0..i).map(|j| 64 << j).collect();
+    let slice: &mut [usize] = &mut data;
+
     let ptr = slice.as_mut_ptr();
     let len = slice.len();
 
     let slice_ptr = slice_ptr_from_raw_parts(ptr, len);
 
     // check that they dereference to the same thing
-    assert_eq!(unsafe { &*slice_ptr }, slice);
+    if unsafe { &*slice_ptr } != slice {
+        return 1;
+    }
     // check that they have the same pointer and length
-    assert_eq!(slice.as_ptr(), slice_ptr.cast::<usize>());
-    assert_eq!(unsafe { slice_ptr.as_ref() }.unwrap().len(), len);
+    if slice.as_ptr() != slice_ptr.cast::<usize>() {
+        return 2;
+    }
+    if unsafe { slice_ptr.as_ref() }.unwrap().len() != len {
+        return 3;
+    }
 
     unsafe {
-        assert_eq!(
-            len,
-            nonnull_slice_len(nonnull_slice_from_raw_parts(
+        if len
+            != nonnull_slice_len(nonnull_slice_from_raw_parts(
                 NonNull::new_unchecked(ptr),
-                len
+                len,
             ))
-        );
+        {
+            return 4;
+        };
     }
 
     for (i, elem) in slice.iter().enumerate() {
         // check that the values are all the same
-        assert_eq!(*elem, unsafe { (&*slice_ptr)[i] });
+        if *elem != unsafe { (&*slice_ptr)[i] } {
+            return 5;
+        }
 
         // manually check that the values are the same
-        assert_eq!(unsafe { (&*slice_ptr)[i] }, 64usize << i);
+        if unsafe { (&*slice_ptr)[i] } != 64usize << i {
+            return 6;
+        }
     }
+
+    0
 }